Mark zeroconf.exceptions as protected by renaming to zeroconf._exceptions
- The public API should only access zeroconf and zeroconf.aio
  as internals may be relocated between releases
