+++ zeroconf/_core.py
@@ -29,6 +29,7 @@ from types import TracebackType  # noqa # used in type hints
 from typing import Dict, List, Optional, Type, Union, cast
 
 from ._dns import DNSIncoming, DNSOutgoing, DNSQuestion
+from ._exceptions import NonUniqueNameException
 from ._handlers import QueryHandler, RecordManager
 from ._logger import QuietLogger, log
 from .cache import DNSCache
@@ -48,7 +49,6 @@ from .const import (
     _TYPE_PTR,
     _UNREGISTER_TIME,
 )
-from .exceptions import NonUniqueNameException
 from .services import (
     RecordUpdateListener,
     ServiceBrowser,
