+++ zeroconf/services/__init__.py
@@ -28,6 +28,7 @@ from collections import OrderedDict
 from typing import Any, Callable, Dict, List, Optional, Set, TYPE_CHECKING, Tuple, Union, cast
 
 from .._dns import DNSAddress, DNSOutgoing, DNSPointer, DNSQuestion, DNSRecord, DNSService, DNSText
+from .._exceptions import BadTypeInNameException
 from ..const import (
     _BROWSER_BACKOFF_LIMIT,
     _BROWSER_TIME,
@@ -47,7 +48,6 @@ from ..const import (
     _TYPE_SRV,
     _TYPE_TXT,
 )
-from ..exceptions import BadTypeInNameException
 from ..utils.name import service_type_name
 from ..utils.net import (
     IPVersion,
